requirements.txt
# Google API Libraries
google-api-python-client==2.92.0   # For interacting with Google APIs, including Google Tasks and Google Calendar
google-auth==2.23.3                # Authentication library for Google APIs
google-auth-oauthlib==1.0.0        # OAuth library for Google's authentication
google-auth-httplib2==0.1.0        # Http library for making authenticated requests

# Pandas - For handling and analyzing data
pandas==1.5.3

# Requests - To make HTTP requests, in case you need additional Taskboard API or other integrations
requests==2.31.0
